+++ src/pages/Main/Main.scss
@@ -1,48 +1,1 @@
-@import '../../styles/variables.scss';
-.mainContainer {
-  width: 1080px;
-  margin: 0px auto 30px;
-  .innerContainer {
-    margin-top: 44px;
-    margin-bottom: 50px;
-  }
 
-  .filterNavBottom {
-    @include flex(space-between, center);
-    height: 40px;
-    padding: 4px 12px;
-    background-color: rgb(246, 246, 246, 0.7);
-  }
-
-  .check {
-    cursor: pointer;
-  }
-
-  .filterButtons {
-    @include flex(flex-end, center);
-    .listFrame {
-      display: flex;
-      .list {
-        padding: 0px 10px;
-      }
-      .btnFrame {
-        outline: none;
-        background: none;
-        appearance: none;
-        border: none;
-        cursor: pointer;
-      }
-    }
-    .filterFrame {
-      cursor: pointer;
-    }
-  }
-
-  .productList {
-    margin-top: 20px;
-    width: 100%;
-    height: auto;
-    @include flex(null, flex-start);
-    flex-wrap: wrap;
-  }
-}
