+++ src/pages/Main/Pagination/Pagination.scss
@@ -1,36 +1,74 @@
-.pagination {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 4px;
-  margin: 16px;
+@import '../../../styles/variables.scss';
 
-  button {
-    border: none;
-    border-radius: 8px;
-    padding: 8px;
-    margin: 0;
-    background: black;
-    color: white;
-    font-size: 1rem;
+.mainContainer {
+  width: 1080px;
+  margin: 0px auto 30px;
+
+  .innerContainer {
+    margin-top: 44px;
+    margin-bottom: 50px;
   }
 
-  &:hover {
-    background: tomato;
+  .filterNavBottom {
+    @include flex(space-between, center);
+    height: 40px;
+    padding: 4px 12px;
+    background-color: rgb(246, 246, 246, 0.7);
+  }
+
+  .check {
     cursor: pointer;
-    transform: translateY(-2px);
   }
 
-  &[disabled] {
-    background: grey;
-    cursor: revert;
-    transform: revert;
+  .filterButtons {
+    @include flex(flex-end, center);
+    .listFrame {
+      display: flex;
+      .list {
+        padding: 0px 10px;
+      }
+      .btnFrame {
+        outline: none;
+        background: none;
+        appearance: none;
+        border: none;
+        cursor: pointer;
+      }
+    }
   }
 
-  &[aria-current] {
-    background: deeppink;
-    font-weight: bold;
-    cursor: revert;
-    transform: revert;
+  .productList {
+    margin-top: 20px;
+    width: 100%;
+    height: auto;
+    @include flex(null, flex-start);
+    flex-wrap: wrap;
+  }
+
+  .page {
+    @include flex(center, center);
+    bottom: 0px;
+    width: 100%;
+    height: 48px;
+    background-color: white;
+    position: fixed;
+    margin: 0 auto;
+    left: 0px;
+  }
+
+  .pageBtn {
+    font-size: 18px;
+    width: 40px;
+    height: 42px;
+    outline: none;
+    background: none;
+    appearance: none;
+    border: none;
+    cursor: pointer;
+  }
+
+  .pageBtn:hover {
+    background-color: rgb(240, 240, 240);
+    border-radius: 50%;
   }
 }
