+++ src/pages/Main/Product/Product.scss
@@ -1,30 +1,28 @@
 .productContainer {
-    width: 200px;
-    margin-left: 20px;
-    margin-bottom: 24px;
-    &:nth-of-type(5n+1) {
-        margin-left: 0px;
+  width: 200px;
+  margin-left: 20px;
+  margin-bottom: 24px;
+  &:nth-of-type(5n + 1) {
+    margin-left: 0px;
+  }
+  .imageContainer {
+    width: 100%;
+    height: 100%;
+    img {
+      width: 100%;
+      height: 200px;
     }
-    .imageContainer {
-        width: 100%;
-        height: 100%;
-        img {
-            width: 100%;
-            height: 200px;
-        }
+  }
+  .productDescription {
+    margin-top: 8px;
+    .name {
+      margin-bottom: 4px;
     }
-    .productDescription {
-        margin-top: 8px;
-        .name {
-            margin-bottom: 4px;
-        }
-        .cost {
-            margin-top: 4px;
-            font-weight: 700;
-            font-size: 18px;
-            line-height: 28px;
-        }
+    .cost {
+      margin-top: 4px;
+      font-weight: 700;
+      font-size: 18px;
+      line-height: 28px;
     }
+  }
 }
-
-
