+++ src/pages/Main/uiData.js
@@ -1,6 +1,4 @@
 export const BUTTON_LIST = [
-  { id: 1, buttonName: '인기상품순' },
-  { id: 2, buttonName: 'kg당 단가순' },
-  { id: 3, buttonName: '개당 단가순' },
-  { id: 4, buttonName: '낮은 가격순' },
+  { id: 1, buttonName: '높은 가격순' },
+  { id: 2, buttonName: '낮은 가격순' },
 ];
